refactor: pertinence --> perimetre
