+++ server/package.json
@@ -30,14 +30,14 @@
     "psup:coverage-2021": "standalone=true node src/jobs/parcoursup/coverage.js",
     "psup:setPsIds-2021": "standalone=true node src/jobs/parcoursup/setPsIdsToFormations.js",
     "psup:reference": "standalone=true node src/jobs/parcoursup/reference/index.js",
-    "psup:pertinence": "standalone=true node src/jobs/parcoursup/pertinence/index.js",
+    "psup:perimetre": "standalone=true node src/jobs/parcoursup/perimetre/index.js",
     "psup:updatemna": "standalone=true node src/jobs/parcoursup/updateMatchInfo.js",
     "psup:resetStatus": "standalone=true node src/jobs/parcoursup/miscJobs/_cleanMnaFStatus.js",
     "psup:report": "standalone=true node src/jobs/parcoursup/reportRejected.js",
     "affelnet:coverage": "standalone=true node src/jobs/affelnet/coverage.js",
     "affelnet:reconciliation": "standalone=true node src/jobs/affelnet/reconciliation.js",
     "affelnet:reference": "standalone=true node src/jobs/affelnet/reference/index.js",
-    "affelnet:pertinence": "standalone=true node src/jobs/affelnet/pertinence/index.js",
+    "affelnet:perimetre": "standalone=true node src/jobs/affelnet/perimetre/index.js",
     "affelnet:import": "standalone=true node src/jobs/oneshot/afFormation/index.js",
     "affelnet:missing-cfd": "standalone=true node src/jobs/oneshot/afFormation/missingCfd.js",
     "es:index": "standalone=true node src/jobs/esIndex/index.js",
