+++ server/src/jobs/parcoursup/perimetre/index.js
@@ -3,22 +3,22 @@ const controller = require("./controller");
 
 const { runScript } = require("../../scriptWrapper");
 
-const afPertinence = async () => {
+const psPerimetre = async () => {
   try {
-    logger.info(" -- Start affelnet pertinence -- ");
+    logger.info(" -- Start psup perimetre -- ");
 
     await controller.run();
 
-    logger.info(" -- End of affelnet pertinence -- ");
+    logger.info(" -- End of psup perimetre -- ");
   } catch (err) {
     logger.error(err);
   }
 };
 
-module.exports = afPertinence;
+module.exports = psPerimetre;
 
 if (process.env.standalone) {
   runScript(async () => {
-    await afPertinence();
+    await psPerimetre();
   });
 }
